refactor(file-structure): separate out utils for raster and vector tiles
