+++ apps/project/graphql/types.py
@@ -10,11 +10,14 @@ from apps.project.project_types.tile_map_service.completeness import project as 
 from apps.project.project_types.tile_map_service.find import project as find_project
 from apps.project.project_types.validate import project as validate_project
 from apps.tutorial.graphql.types import TutorialType
-from utils.geo.tile_server.config import TileServerNameEnum, VectorTileServerNameEnum
+from utils.geo.tile_server.config import TileServerNameEnum
 from utils.geo.tile_server.models import (
     TileServerCommonConfig,
     TileServerConfig,
     TileServerCustomConfig,
+)
+from utils.geo.vector_tile_server.config import VectorTileServerNameEnum
+from utils.geo.vector_tile_server.models import (
     VectorTileServerCommonConfig,
     VectorTileServerConfig,
     VectorTileServerCustomConfig,
