+++ utils/geo/tile_server/config.py
@@ -2,13 +2,6 @@ from django.conf import settings
 from django.db import models
 
 
-class VectorTileServerNameEnum(models.TextChoices):
-    CUSTOM = "CUSTOM", "Custom"
-    OPEN_STREET_MAP = "OPEN_STREET_MAP", "Open Street Map"
-    OPEN_FREE_MAP = "OPEN_FREE_MAP", "Open Free Map"
-    VERSATILES = "VERSATILES", "Versatiles"
-
-
 class TileServerNameEnum(models.TextChoices):
     CUSTOM = "CUSTOM", "Custom"
     BING = "BING", "Bing"
@@ -21,15 +14,6 @@ class TileServerNameEnum(models.TextChoices):
 
 
 class Config:
-    VECTOR_IMAGE_URLS = {
-        # VectorTileServerNameEnum.OPEN_STREET_MAP: "https://vector.osm.org/demo/shortbread/colorful.json",
-        VectorTileServerNameEnum.OPEN_STREET_MAP: "https://vector.osm.org/shortbread_v1/{z}/{x}/{y}.mvt",
-        # VectorTileServerNameEnum.OPEN_FREE_MAP: "https://tiles.openfreemap.org/styles/liberty",
-        VectorTileServerNameEnum.OPEN_FREE_MAP: "https://tiles.openfreemap.org/planet/20250528_001001_pt/{z}/{x}/{y}.pbf",
-        # VectorTileServerNameEnum.VERSATILES: "https://tiles.versatiles.org/assets/styles/colorful/style.json",
-        VectorTileServerNameEnum.VERSATILES: "https://tiles.versatiles.org/tiles/osm/{z}/{x}/{y}",
-    }
-
     # FIXME(tnagorra): This will be obsolete soon
     IMAGE_URLS = {
         TileServerNameEnum.BING: ("https://ecn.t0.tiles.virtualearth.net/tiles/a{quadkey}.jpeg?g=7505&token={key}"),
